pyproject.toml
[project]
name = "streamlit-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.4.1",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "scikit-learn>=1.8.0",
    "streamlit>=1.53.1",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
]
